+++ src/main/java/tornadofx/Dialogs.kt
@@ -14,8 +14,8 @@ import java.io.File
  * Show a confirmation dialog and execute the given action if confirmButton is clicked. The button types
  * of the confirmButton and cancelButton are configurable.
  */
-fun confirm(header: String, content: String = "", confirmButton: ButtonType = ButtonType.OK, cancelButton: ButtonType = ButtonType.CANCEL, actionFn: () -> Unit) {
-    alert(Alert.AlertType.CONFIRMATION, header, content, *arrayOf(confirmButton, cancelButton)) {
+fun confirm(header: String, content: String = "", confirmButton: ButtonType = ButtonType.OK, cancelButton: ButtonType = ButtonType.CANCEL, owner: Window? = null, title: String? = null, actionFn: () -> Unit) {
+    alert(Alert.AlertType.CONFIRMATION, header, content, confirmButton, cancelButton, owner = owner, title = title) {
         if (it == confirmButton) actionFn()
     }
 }
@@ -33,9 +33,11 @@ fun alert(type: Alert.AlertType,
           content: String? = null,
           vararg buttons: ButtonType,
           owner: Window? = null,
+          title: String? = null,
           actionFn: (Alert.(ButtonType) -> Unit)? = null): Alert {
 
     val alert = Alert(type, content ?: "", *buttons)
+    title?.let { alert.title = it }
     alert.headerText = header
     owner?.also { alert.initOwner(it) }
     val buttonClicked = alert.showAndWait()
@@ -43,17 +45,17 @@ fun alert(type: Alert.AlertType,
     return alert
 }
 
-fun warning(header: String, content: String? = null, vararg buttons: ButtonType, actionFn: (Alert.(ButtonType) -> Unit)? = null) =
-        alert(Alert.AlertType.WARNING, header, content, *buttons, actionFn = actionFn)
+fun warning(header: String, content: String? = null, vararg buttons: ButtonType, owner: Window? = null, title: String? = null, actionFn: (Alert.(ButtonType) -> Unit)? = null) =
+        alert(Alert.AlertType.WARNING, header, content, *buttons, owner = owner, title = title, actionFn = actionFn)
 
-fun error(header: String, content: String? = null, vararg buttons: ButtonType, actionFn: (Alert.(ButtonType) -> Unit)? = null) =
-        alert(Alert.AlertType.ERROR, header, content, *buttons, actionFn = actionFn)
+fun error(header: String, content: String? = null, vararg buttons: ButtonType, owner: Window? = null, title: String? = null, actionFn: (Alert.(ButtonType) -> Unit)? = null) =
+        alert(Alert.AlertType.ERROR, header, content, *buttons, owner = owner, title = title, actionFn = actionFn)
 
-fun information(header: String, content: String? = null, vararg buttons: ButtonType, actionFn: (Alert.(ButtonType) -> Unit)? = null) =
-        alert(Alert.AlertType.INFORMATION, header, content, *buttons, actionFn = actionFn)
+fun information(header: String, content: String? = null, vararg buttons: ButtonType, owner: Window? = null, title: String? = null, actionFn: (Alert.(ButtonType) -> Unit)? = null) =
+        alert(Alert.AlertType.INFORMATION, header, content, *buttons, owner = owner, title = title, actionFn = actionFn)
 
-fun confirmation(header: String, content: String? = null, vararg buttons: ButtonType, actionFn: (Alert.(ButtonType) -> Unit)? = null) =
-        alert(Alert.AlertType.CONFIRMATION, header, content, *buttons, actionFn = actionFn)
+fun confirmation(header: String, content: String? = null, vararg buttons: ButtonType, owner: Window? = null, title: String? = null, actionFn: (Alert.(ButtonType) -> Unit)? = null) =
+        alert(Alert.AlertType.CONFIRMATION, header, content, *buttons, owner = owner, title = title, actionFn = actionFn)
 
 enum class FileChooserMode { None, Single, Multi, Save }